+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * ignore output from executor and show only stdout/stderr from actual command (#11)
 * return exit code 1 if at least one task failed (#12)
 * allow selecting slaves with constraints (#6)
+* allow to specify mesos role (#5)
 
 ## v0.1.0